Invent Buttons
Fixed which ones should and shouldn't be red

+++ src/main/java/core/Decoder.java
@@ -59,7 +59,7 @@ public class Decoder {
         String[] data = DataHandler.readFile(DataHandler.dir + "//player.gme");
         double x = Double.parseDouble(data[1]);
         double y = Double.parseDouble(data[2]);
-        int health = Integer.parseInt(data[3]);
+        double health = Double.parseDouble(data[3]);
         double xOffset = Double.parseDouble(data[4]);
         
         player = new Player(map, new Vector(x, y), "/player.png", xOffset);
